Update scripts to run tests
Fix restores source mapping for tests to make stacktrace to
report actual line:column for test ts file whhere it failed.
It also gets read of index files for running tests in duebug
and coverage mode.
