+++ package.json
@@ -1037,6 +1037,7 @@
     "rxjs": "^6.5.3",
     "semver": "^6.3.0",
     "shelljs": "^0.8.3",
+    "source-map-support": "^0.5.16",
     "string-format": "^2.0.0",
     "targz": "^1.0.1",
     "unzip-stream": "^0.3.0",
